% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-functions.R
\name{add_pi_poly}
\alias{add_pi_poly}
\title{Add a prediction interval}
\usage{
add_pi_poly(plot, data, x, newcat, model, level = 0.95,
  one_line = FALSE)
}
\arguments{
\item{plot}{a ggplot}

\item{data}{data table}

\item{x}{string of x variable name}

\item{newcat}{categorical variable vector}

\item{model}{created lm model}

\item{level}{level of the interval}

\item{one_line}{boolean. Set to TRUE if using rl_poly_same_line, otherwise FALSE}
}
\value{
a ggplot
}
\description{
Add prediction interval to ggplot with specific interval width
}
