% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-functions.R
\name{ggLinearModel}
\alias{ggLinearModel}
\title{ggLinearModel}
\usage{
ggLinearModel(data, x, y, cat = NULL, plotly = FALSE,
  same_slope = FALSE, same_intercept = FALSE, poly = 1,
  interactions = poly, ci = FALSE, pi = FALSE, interactive = FALSE,
  title = NULL, xlabel = NULL, ylabel = NULL, legendTitle = NULL,
  level = 0.95, coefficients = FALSE, model = FALSE)
}
\arguments{
\item{data}{a R data frame. tibbles will work}

\item{x}{variable name of the x (predicting) variable, must be a column name in the data table}

\item{y}{variable name of the y (response) variable, must be a column name in the data table}

\item{cat}{variable name of the categorical variable, must be a column name in the data table. Set to NULL}

\item{plotly}{boolean if they want the graph in plotly. TRUE returns plotly graph, FALSE returns ggplot. Set to FALSE}

\item{same_slope}{boolean if they want the regression lines to have the same slope. Set to FALSE}

\item{same_intercept}{boolean if they want the regression lines to have the same intercept Set to FALSE}

\item{poly}{a integer value to declare how many polynomial variables the use wants. Set to 1, i.e. no polynomial}

\item{interactions}{a integer value for if the user wants to include interactions in the polynomial model. Note: only matters when poly > 1 and interactions <= poly. Set to 0}

\item{ci}{boolean to add a confidence interval to the graph. Set to FALSE}

\item{pi}{boolean to add a prediction interval to the graph. Set to FALSE}

\item{interactive}{boolean to make interactive. Set to FALSE}

\item{title}{a string to change the title of the graph. Set to x name vs y name}

\item{xlabel}{a string for the label of the x axis. Set to the x variable name}

\item{ylabel}{a string for the label of the y axis. Set to the y variable name}

\item{legendTitle}{a string for the label of the legend. Set to the categorical variable name}

\item{level}{a number between 0-1 for the interval percentage of the confidence and prediction intervals. Set to .95}

\item{coefficients}{a boolean to print the coefficients of the model. Set to FALSE}

\item{model}{a boolean to return the model used to create the LSRL. Set to FALSE}
}
\value{
a ggplot or plotly object, and a model object if requested
}
\description{
\pkg{ggLinearModel()} creates a ggplot (or plotly) object. It is designed to create visualizations of various types of linear models. Inserting a datatable, x variable, y variable, and categorical variable will produce a graph with a different least squares line for each level in the categorical variable. You can modify the graph by changing the values for various parameters.
}
\examples{
ggLinearModel(iris, Sepal.Length, Sepal.Width)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, ci = TRUE)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, pi = TRUE)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, ci = TRUE, pi = TRUE, level = .5)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, title = 'Sepal Width based on Sepal Width and Species')
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, model = TRUE)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, poly = 2, interactions = 1)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, coefficients = TRUE)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, plotly = TRUE)
ggLinearModel(iris, Sepal.Length, Sepal.Width, Species, interactive = TRUE)
}
